requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
